requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.6
django-stdimage==6.0.1
gunicorn==21.2.0
packaging==23.2
Pillow==10.0.1
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.8.0
